Reward modeling init
