+++ src/lmflow/pipeline/utils/reward_dataprocessor.py
@@ -1,0 +1,73 @@
+from dataclasses import dataclass
+from typing import Any, Dict, List, Optional, Union
+
+from datasets import load_dataset
+from transformers import AutoTokenizer
+from transformers.utils import PaddingStrategy
+
+
+@dataclass
+class RewardDataCollatorWithPadding:
+    tokenizer: AutoTokenizer
+    padding: Union[bool, str, PaddingStrategy] = True
+    max_length: Optional[int] = None
+    pad_to_multiple_of: Optional[int] = None
+    return_tensors: str = "pt"
+
+    def __call__(self, features: List[Dict[str, Any]]) -> Dict[str, Any]:
+        merged_features = []
+        for feature in features:
+            merged_features.append(
+                {
+                    "input_ids": feature["input_ids_j"],
+                    "attention_mask": feature["attention_mask_j"],
+                }
+            )
+            merged_features.append(
+                {
+                    "input_ids": feature["input_ids_k"],
+                    "attention_mask": feature["attention_mask_k"],
+                }
+            )
+        batch = self.tokenizer.pad(
+            merged_features,
+            padding=self.padding,
+            max_length=self.max_length,
+            pad_to_multiple_of=self.pad_to_multiple_of,
+            return_tensors=self.return_tensors,
+        )
+        batch = {
+            "input_ids": batch["input_ids"],
+            "attention_mask": batch["attention_mask"],
+            "return_loss": True,
+        }
+        return batch
+    
+    
+
+def build_dataset(tokenizer, train_path, eval_path):
+
+    def tokenize(sample):
+        
+        sample['positive'] = tokenizer.apply_chat_template(
+            sample['chosen'], tokenize=False, add_generation_prompt=False).replace(tokenizer.bos_token, "")
+        sample['negative'] = tokenizer.apply_chat_template(
+            sample['rejected'], tokenize=False, add_generation_prompt=False).replace(tokenizer.bos_token, "")
+        
+        tokenized_pos = tokenizer(sample['positive'], truncation=True)
+        tokenized_neg = tokenizer(sample['negative'], truncation=True)
+        sample["input_ids_j"] = tokenized_pos["input_ids"]
+        sample["attention_mask_j"] = tokenized_pos["attention_mask"]
+        sample["input_ids_k"] = tokenized_neg["input_ids"]
+        sample["attention_mask_k"] = tokenized_neg["attention_mask"]
+        return sample
+    
+    ds = load_dataset(train_path, split="train").shuffle(seed=42)
+    ds = ds.map(tokenize, num_proc=8)
+
+    eval_dataset = None
+
+    train_dataset = ds
+    eval_dataset = load_dataset(eval_path, split="train").shuffle(seed=42).select(range(500))
+    
+    return train_dataset, eval_dataset
